src/ai/main.py
"""
AI Realtor Assistant - Main Application
"""

from fastapi import FastAPI, HTTPException, Depends
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from fastapi.middleware.cors import CORSMiddleware
from typing import Optional
from datetime import datetime, timedelta
from jose import JWTError, jwt
from passlib.context import CryptContext
from pydantic import BaseModel
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="AI Realtor Assistant",
    description="AI-powered automation tools for real estate professionals",
    version="1.0.0"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # In production, replace with specific origins
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Security
SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key")
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")

# Models
class Token(BaseModel):
    access_token: str
    token_type: str

class TokenData(BaseModel):
    username: Optional[str] = None

class User(BaseModel):
    username: str
    email: Optional[str] = None
    full_name: Optional[str] = None
    subscription_tier: str = "basic"  # basic, pro, premium
    disabled: Optional[bool] = None

class SubscriptionPlan(BaseModel):
    name: str
    price: float
    features: list[str]

# Subscription plans
SUBSCRIPTION_PLANS = {
    "basic": SubscriptionPlan(
        name="Basic",
        price=49.0,
        features=[
            "AI chatbot for lead qualification",
            "Basic lead management",
            "Email support"
        ]
    ),
    "pro": SubscriptionPlan(
        name="Pro",
        price=149.0,
        features=[
            "Everything in Basic",
            "AI scheduling assistant",
            "Automated email follow-ups",
            "Social media posting",
            "Priority support"
        ]
    ),
    "premium": SubscriptionPlan(
        name="Premium",
        price=299.0,
        features=[
            "Everything in Pro",
            "AI CRM integration",
            "Contract automation",
            "Market analysis",
            "Advanced prospecting",
            "24/7 support"
        ]
    )
}

# Security functions
def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password):
    return pwd_context.hash(password)

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

# Routes
@app.get("/")
async def root():
    return {"message": "Welcome to AI Realtor Assistant API"}

@app.get("/plans")
async def get_subscription_plans():
    return SUBSCRIPTION_PLANS

@app.post("/token", response_model=Token)
async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
    # TODO: Implement actual user authentication
    user = authenticate_user(form_data.username, form_data.password)
    if not user:
        raise HTTPException(
            status_code=401,
            detail="Incorrect username or password",
            headers={"WWW-Authenticate": "Bearer"},
        )
    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = create_access_token(
        data={"sub": user.username}, expires_delta=access_token_expires
    )
    return {"access_token": access_token, "token_type": "bearer"}

# Placeholder functions
def authenticate_user(username: str, password: str):
    # TODO: Implement actual user authentication
    return {"username": username, "subscription_tier": "basic"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000) 