src/ai/requirements.txt
# Core AI Dependencies
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=0.19.0

# Communication & Integration
twilio>=8.10.0
zapier-python>=0.1.0
calendly>=0.1.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.1.0

# Real Estate APIs
zillow-api>=0.1.0
realtor-api>=0.1.0

# Social Media & Marketing
canva-api>=0.1.0
hootsuite-api>=0.1.0
buffer-api>=0.1.0

# Document Processing
docusign-esign>=3.20.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# CRM Integration
hubspot-api-client>=8.0.0
zoho-crm>=0.1.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=4.1.0
aiohttp>=3.8.0
pydantic>=2.0.0 